Created second version of the scan system.
In this version, I defined three tags in a list, seen at line 4. I was able to scan for the appearance within the 'Files' directory of all three tags simultaneously.
Because I gave the destination directories the same name as the tags, an error came up where the program tried to move the directories into themselves. This was fixed using the operation at line 14.